tradeinfovector/requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
async-timeout==5.0.1
beautifulsoup4==4.14.3
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dnspython==2.8.0
fastapi==0.135.3
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
kombu==5.6.2
motor==3.7.1
packaging==26.0
playwright==1.58.0
prompt_toolkit==3.0.52
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
redis==6.4.0
six==1.17.0
soupsieve==2.8.3
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
uvicorn==0.44.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
